add performanceCard to modules
